+++ packages/uni-app-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @js-css/uni-app-react
 
+## 0.0.13
+
+### Patch Changes
+
+- fix: number 0 text context update to ''
+
 ## 0.0.12
 
 ### Patch Changes